backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
aiosqlite>=0.18.0
python-multipart>=0.0.6
exifread>=3.0.0
PyPDF2>=3.0.0
mutagen>=1.46.0
python-docx>=0.8.11

# Для визначення типів файлів
python-magic>=0.4.27; platform_system != "Windows"  # Для Linux/macOS
python-magic-bin>=0.4.14; platform_system == "Windows"  # Для Windows